refactor: Update name of controllers and controller folder
controller 폴더명을 api로 변경
xxxController 클래스의 이름을 xxxApiController로 변경
