+++ src/main/java/jandy3/DesignSeller/api/MarketApiController.java
@@ -1,11 +1,10 @@
-package jandy3.DesignSeller.controller;
+package jandy3.DesignSeller.api;
 
 import jandy3.DesignSeller.auth.PrincipalDetails;
 import jandy3.DesignSeller.dto.IdResponse;
 import jandy3.DesignSeller.oauth.annotation.CurrentUser;
 import jandy3.DesignSeller.domain.Market;
 import jandy3.DesignSeller.dto.MarketRequest;
-import jandy3.DesignSeller.paylaod.ApiResponse;
 import jandy3.DesignSeller.service.MarketService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -15,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-public class MarketController {
+public class MarketApiController {
 
     private final MarketService marketService;
